refactor(web-server): modularize shared-routes and Windows-aware symlink status
